Add exception control

+++ src/Controller/PostsController.php
@@ -1,6 +1,10 @@
 <?php
 namespace App\Controller;
+
 use Cake\View\JsonView;
+use Cake\Http\Exception\NotFoundException;
+use Cake\Http\Exception\BadRequestException;
+use Cake\Http\Exception\InternalErrorException;
 
 class PostsController extends AppController
 {
@@ -12,6 +16,9 @@ class PostsController extends AppController
     public function index()
     {
         $posts = $this->Posts->find('all')->all();
+        if (!$posts) {
+            throw new NotFoundException(__('Posts not found'));
+        }
         $this->set('posts', $posts);
         $this->viewBuilder()->setOption('serialize', ['posts']);
     }
@@ -19,6 +26,9 @@ class PostsController extends AppController
     public function view($id)
     {
         $post = $this->Posts->get($id);
+        if (!$post) {
+            throw new NotFoundException(__('Post not found'));
+        }
         $this->set('post', $post);
         $this->viewBuilder()->setOption('serialize', ['post']);
     }
@@ -28,7 +38,11 @@ class PostsController extends AppController
         $this->request->allowMethod(['post']);
         $post = $this->Posts->newEntity($this->request->getData());
 
-        $message =  $this->Posts->save($post) ?'Saved' : 'Error';
+        if (!$post->getErrors() && $this->Posts->save($post)) {
+            $message = 'Saved';
+        } else {
+            throw new BadRequestException(__('Unable to save post'));
+        }
 
         $this->set([
             'message' => $message,
@@ -42,11 +56,15 @@ class PostsController extends AppController
         $this->request->allowMethod(['delete']);
 
         $post = $this->Posts->get($id);
-        $message = 'Deleted';
-        if (!$this->Posts->delete($post)) {
-            $message = 'Error';
+        if (!$post) {
+            throw new NotFoundException(__('Post not found'));
         }
-        $this->set('message', $message);
+
+        if (!$this->Posts->delete($post)) {
+            throw new InternalErrorException(__('Unable to delete post'));
+        }
+
+        $this->set('message', 'Deleted');
         $this->viewBuilder()->setOption('serialize', ['message']);
     }
 }
